fix: Reconfigure Cobra configuration
In the past, the application was instantiating Cobra configuration using
`init` blocks which is considered poor practice these days.

This change restructures the application code to make it easy to have a
monobinary where each application run mode is decoupled from the others.

This also fixes the problem where in the past, running `go run main.go`
without any command would be a noop action. This restructure makes it
such that if no command is specified, Cobra defaults to the `help`
command and prints out help docs.

```
⇒  go run cmd/cloudwaste.go
Cloudwaste finds wasted resources in your cloud

Usage:
  cloudwaste [command]

Available Commands:
  help        Help about any command
  scan        Scan your cloud accounts for unused resources

Flags:
  -h, --help   help for cloudwaste

Use "cloudwaste [command] --help" for more information about a command.
```
