+++ .goreleaser.yml
@@ -15,6 +15,6 @@ builds:
       - windows
     goarch:
       - amd64
-    main: ./main.go
+    main: ./cmd/cloudwaste.go
     binary: cloudwaste
     id: "build-cloudwaste"
